Added validation for overlapping policy range and checks for deletion of policy if it is in use

+++ src/main/java/com/reservation/service/CancellationPolicyService.java
@@ -5,16 +5,28 @@ import com.reservation.dto.CancellationPolicyResponse;
 import com.reservation.entity.CancellationPolicy;
 import com.reservation.mapper.Mapper;
 import com.reservation.repository.CancellationPolicyRepository;
+import com.reservation.repository.CancellationRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpStatusCodeException;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor
 public class CancellationPolicyService {
     private final CancellationPolicyRepository cancellationPolicyRepository;
+    private final CancellationRepository cancellationRepository;
 
     public CancellationPolicyResponse createCancellationPolicy(CancellationPolicyRequest policy){
         CancellationPolicy cancellationPolicy = new CancellationPolicy();
+
+        boolean exists = cancellationPolicyRepository.existsOverlappingRange(policy.getDaysBeforeCheckInFrom(), policy.getDaysBeforeCheckInTo(), null);
+        if (exists)
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST,
+                    "Overlapping policy date range"
+            );
 
         Mapper.mapRequestToEntity(cancellationPolicy , policy);
 
@@ -27,6 +39,13 @@ public class CancellationPolicyService {
         CancellationPolicy cancellationPolicy = cancellationPolicyRepository.findById(id)
                         .orElseThrow(()-> new RuntimeException("Invalid Policy"));
 
+        boolean exists = cancellationPolicyRepository.existsOverlappingRange(policy.getDaysBeforeCheckInFrom(), policy.getDaysBeforeCheckInTo(), id);
+        if (exists)
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST,
+                    "Overlapping policy date range"
+            );
+
         Mapper.mapRequestToEntity(cancellationPolicy , policy);
 
         CancellationPolicy savedPolicy = cancellationPolicyRepository.save(cancellationPolicy);
@@ -35,6 +54,14 @@ public class CancellationPolicyService {
     }
 
     public void deleteCancellationPolicy(Integer id){
+        boolean exists = cancellationRepository.existsByCancellationPolicy_Id(id);
+
+        if (exists)
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST,
+                    "Policy is in use and cannot be deleted"
+            );
+
         cancellationPolicyRepository.deleteById(id);
     }
